requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.2,<4.0
celery>=5.4,<6.0
redis>=5.0,<6.0
python-dotenv>=1.0,<2.0
pgvector>=0.3,<0.4
pypdf>=5.0.0,<6.0.0
openai>=1.0.0,<2.0.0
